small fixes and restructering of project
