+++ setup.cfg
@@ -10,4 +10,4 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    pgn-download = pgn_downloader.main:cli_entrypoint
+    pgn-downloader = pgn_downloader.__main__:main
